requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
sqlalchemy>=2.0
pydantic-settings>=2.2
python-dotenv>=1.0
passlib>=1.7
python-multipart>=0.0.9
jinja2>=3.1
itsdangerous>=2.1
httpx>=0.27
pypdf>=4.0
pillow>=10.0
pytesseract>=0.3.10
sentence-transformers>=3.0.0
